+++ src/main/java/com/revolvingmadness/sculk/language/builtins/classes/types/IntegerType.java
@@ -88,8 +88,16 @@ public class IntegerType extends BuiltinType {
             BuiltinClass other = arguments.get(0);
 
             if (other.instanceOf(new FloatType())) {
+                if (other.toFloat() == 0.0) {
+                    throw ErrorHolder.cannotDivideByZero();
+                }
+
                 return new FloatInstance(this.boundClass.toInteger() / other.toFloat());
             } else if (other.instanceOf(new IntegerType())) {
+                if (other.toInteger() == 0) {
+                    throw ErrorHolder.cannotDivideByZero();
+                }
+
                 return new IntegerInstance(this.boundClass.toInteger() / other.toInteger());
             }