+++ src/main/java/com/revolvingmadness/sculk/language/errors/DivisionByZeroError.java
@@ -1,0 +1,7 @@
+package com.revolvingmadness.sculk.language.errors;
+
+public class DivisionByZeroError extends Error {
+    public DivisionByZeroError() {
+        super("Cannot divide by zero");
+    }
+}